ADDED AUTHENTICATION WITH MVVM
