+++ app/src/main/res/layout/fragment_login.xml
@@ -7,7 +7,7 @@
     android:background="@color/App_theme_color"
     android:fillViewport="true"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".LoginFragment">
+    tools:context=".view.LoginFragment">
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
